+++ packages/server/src/emitters/roomEmitter.ts
@@ -1,6 +1,6 @@
 import { Server as IOServer } from 'socket.io';
 import { RoomEvent } from 'shared';
-import { getRooms } from '../controllers/roomControllerMDB';
+import { getRooms } from '../controllers/roomController';
 
 export async function emitRoomUpdate(io: IOServer) {
   const rooms = await getRooms();
